Add serverless cloud function namespace automatic lookup.
Add the namespace automatic search function of serverless cloud functions. If a region unique ID or namespace name is not given when calling a serverless cloud function, the relevant API will try to automatically find and use the unique ID and namespace name of the region where the serverless cloud function is located. It is worth noting that the default automatic search function is only applicable to the use of related APIs in serverless cloud function products.